pyproject.toml
[project]
name = "horse-racing-ml"
version = "0.1.0"
description = "Horse racing prediction ML system on GCP"
requires-python = ">=3.10"
dependencies = [
    "polars>=1.0.0",
    "pandas>=2.0.0",
    "lightgbm>=4.0.0",
    "scikit-learn>=1.3.0",
    "google-cloud-storage>=2.14.0",
    "google-cloud-bigquery>=3.14.0",
    "google-cloud-secret-manager>=2.18.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0",
    "mlflow>=2.10.0",
    "optuna>=3.5.0",
    "structlog>=24.1.0",
    "httpx>=0.27.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.4.0",
    "black>=24.0.0",
    "mypy>=1.8.0",
    "types-PyYAML>=6.0",
]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = [
    "N803",  # X (uppercase) is ML convention for feature matrices
    "N806",  # X_train, X_test are ML convention
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["E501"]  # Allow longer lines in tests

[tool.black]
target-version = ["py310"]
line-length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["sklearn.*", "lightgbm.*", "optuna.*", "mlflow.*", "structlog.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
# Polars .item() returns broad union types - relax checks for modules using it heavily
module = [
    "src.evaluator.monitor",
    "src.evaluator.backtest_engine",
    "src.evaluator.metrics",
    "src.feature_engineering.transformers.scalers",
]
disallow_untyped_defs = true
warn_return_any = false
disable_error_code = ["arg-type", "operator"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pandas-stubs>=2.3.3.260113",
]
